+++ __tests__/crypto.js
@@ -204,8 +204,15 @@ test('crypto', async () => {
     expect(ton_public).toEqual('PubDdJkMyss2qHywFuVP1vzww0TpsLxnRNnbifTCcu-XEgW0');
 
 
-    const phrase = await crypto.mnemonicFromRandom();
+    let phrase = await crypto.mnemonicFromRandom();
     expect(phrase.split(' ').length).toEqual(24);
+
+
+    phrase = await crypto.mnemonicFromRandom({ dictionary: 0, wordCount: 12});
+    expect(phrase.split(' ').length).toEqual(12);
+
+    phrase = await crypto.mnemonicFromRandom({ dictionary: 1, wordCount: 12});
+    expect(phrase.split(' ').length).toEqual(12);
 
 
     const entropy = '2199ebe996f14d9e4e2595113ad1e6276bd05e2e147e16c8ab8ad5d47d13b44fcf';